refactor(tests): #58 Tier 1.2 — split validation framework tests by layer
tests/test_validation_types.py (1,097 LOC, 78 tests, 16 Test* classes) →
tests/validation/{core,data,completeness,lineage}.py.

  test_core.py          4 classes, 20 tests — CheckResult / ValidationReport
                                              / ValidationStore /
                                              ValidationEngine framework
                                              types — 265 LOC
  test_data.py          3 classes, 13 tests — schema / series / anomaly
                                              detection — 192 LOC
  test_completeness.py  5 classes, 23 tests — catalog completeness /
                                              cross-source / data diff /
                                              volume / freshness — 320 LOC
  test_lineage.py       3 classes, 22 tests — revision monitoring /
                                              lineage / dimension — 327 LOC

Each split file imports only the validation symbols it references and
keeps the original `sys.path.insert(0, ...src)` shim so it can run
standalone outside ``PYTHONPATH=src``. Test bodies are byte-equivalent.

Verification: 78 tests pass on branch (same as master baseline).
